+++ src/main/java/com/example/tradedemo/domain/marketlistings/service/MarketListingService.java
@@ -108,8 +108,6 @@ public class MarketListingService {
     public PageResponse<SearchAllMarketListingResponse> getAllMarketListing(
             Long memberId, String keyword, String sortTotalPrice, String sortSaleEndAt, Pageable pageable) {
 
-        expireMarketListings();   // 만료 처리
-
         if (keyword != null && !keyword.isBlank()) {
             marketListingCacheService.cacheSearchKeyword(memberId, keyword);
         }
@@ -135,27 +133,7 @@ public class MarketListingService {
                 .findById(marketListingId)
                 .orElseThrow(() -> new ServiceException(ErrorEnum.ERR_MARKET_LISTING_NOT_FOUND));
 
-        expireMarketListings();
-
         return SearchMarketListingResponse.of(marketListing);
-    }
-    /**
-     * 만료 시간
-     * 상품 등록 시 만료 시간 체크 : saleEndAt
-     * 만료 시간이 되면 SELLING(판매) → EXPIRED(만료)
-     */
-    @Transactional
-    public void expireMarketListings() {
-
-        List<MarketListing> expiredListings =
-                marketListingRepository.findByStatusAndSaleEndAtBefore(
-                        MarketListingStatus.SELLING,
-                        LocalDateTime.now()
-                );
-
-        for (MarketListing listing : expiredListings) {
-            listing.updateStatus(MarketListingStatus.EXPIRED);
-        }
     }
 
     @Transactional(readOnly = true)
